+++ src/main.rs
@@ -34,8 +34,8 @@ fn main() -> Result<(), Box<dyn Error>> {
     puzzle(1, 'a', Box::new(util::get_parsed_lines::<u32>), Box::new(day01::day01a))?;
     puzzle(1, 'b', Box::new(util::get_parsed_lines::<u32>), Box::new(day01::day01b))?;
 
-    puzzle(2, 'a', Box::new(util::get_splitted_commas_numbers::<usize>), Box::new(day02::day02a))?;
-    puzzle(2, 'b', Box::new(util::get_splitted_commas_numbers::<usize>), Box::new(day02::day02b))?;
+    puzzle(2, 'a', Box::new(util::get_splitted_commas_numbers::<i128>), Box::new(day02::day02a))?;
+    puzzle(2, 'b', Box::new(util::get_splitted_commas_numbers::<i128>), Box::new(day02::day02b))?;
 
     puzzle(3, 'a', Box::new(util::get_parsed_lines::<day03::Input>), Box::new(day03::day03a))?;
     puzzle(3, 'b', Box::new(util::get_parsed_lines::<day03::Input>), Box::new(day03::day03b))?;
